+++ pkg/skaffold/render/generate/generate_test.go
@@ -63,57 +63,6 @@ spec:
   - name: leeroy-web3
     image: leeroy-web3
 `
-
-	// Test file under <tmp>/base/patch.yaml
-	patchYaml = `apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: kustomize-test
-spec:
-  template:
-    spec:
-      containers:
-      - name: kustomize-test
-        image: index.docker.io/library/busybox
-        command:
-          - sleep
-          - "3600"
-`
-	// Test file under <tmp>/base/deployment.yaml
-	kustomizeDeploymentYaml = `apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: kustomize-test
-  labels:
-    app: kustomize-test
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: kustomize-test
-  template:
-    metadata:
-      labels:
-        app: kustomize-test
-    spec:
-      containers:
-      - name: kustomize-test
-        image: not/a/valid/image
-`
-
-	// Test file under <tmp>/base/kustomization.yaml
-	kustomizeYaml = `resources:
-  - deployment.yaml
-patches:
-  - patch.yaml
-`
-	// Test file under <tmp>/fn/Kptfile
-	kptfileYaml = `apiVersion: kpt.dev/v1
-kind: Kptfile
-metadata:
-  name: fake-fn
-pipeline:
-`
 )
 
 func TestGenerate(t *testing.T) {
@@ -138,38 +87,6 @@ func TestGenerate(t *testing.T) {
 			},
 			expected: manifest.ManifestList{[]byte(podYaml), []byte(podsYaml)},
 		},
-		{
-			description: "render kustomize manifests",
-			generateConfig: latest.Generate{
-				Kustomize: &latest.Kustomize{
-					Paths: []string{"base"},
-				},
-			},
-			commands: testutil.CmdRunOut("kustomize build base", podsYaml),
-			expected: manifest.ManifestList{[]byte(podsYaml)},
-		},
-		{
-			description: "render kustomize manifests - kubectl",
-			generateConfig: latest.Generate{
-				Kustomize: &latest.Kustomize{
-					Paths: []string{"base"},
-				},
-			},
-			useKubectlKustomize: true,
-			commands:            testutil.CmdRunOut("kustomize build base", patchYaml),
-			expected:            manifest.ManifestList{[]byte(patchYaml)},
-		},
-
-		{
-			description: "render kpt manifests",
-			generateConfig: latest.Generate{
-				Kpt: []string{"Kptfile"},
-			},
-			// Using "filepath" to join path so as the result can fix when running in either linux or
-			// windows (skaffold integration test).
-			commands: testutil.CmdRun(fmt.Sprintf("kpt fn render fn --output=%v", "fn")),
-			expected: manifest.ManifestList{},
-		},
 	}
 	for _, test := range tests {
 		testutil.Run(t, test.description, func(t *testutil.T) {
@@ -183,17 +100,12 @@ func TestGenerate(t *testing.T) {
 			t.NewTempDir().
 				Write("pod.yaml", podYaml).
 				Write("pods.yaml", podsYaml).
-				Write("base/kustomization.yaml", kustomizeYaml).
-				Write("base/patch.yaml", patchYaml).
-				Write("base/deployment.yaml", kustomizeDeploymentYaml).
-				Write("fn/Kptfile", kptfileYaml).
 				Touch("empty.ignored").
 				Chdir()
 
 			g := NewGenerator(".", test.generateConfig, "")
 			var output bytes.Buffer
 			actual, err := g.Generate(context.Background(), &output)
-			defer os.RemoveAll(".kpt-pipeline")
 			t.CheckNoError(err)
 			t.CheckDeepEqual(actual.String(), test.expected.String())
 		})
@@ -240,37 +152,13 @@ func TestManifestDeps(t *testing.T) {
 			expected: []string{"rawYaml-sample/pod.yaml"},
 		},
 		{
-			description: "kustomize dir",
-			generateConfig: latest.Generate{
-				Kustomize: &latest.Kustomize{
-					Paths: []string{"kustomize-sample"},
-				},
-			},
-			expected: []string{"kustomize-sample/kustomization.yaml", "kustomize-sample/patch.yaml"},
-		},
-		{
-			description: "kpt dir",
-			generateConfig: latest.Generate{
-				Kpt: []string{"kpt-sample"},
-			},
-			expected: []string{"kpt-sample/Kptfile", "kpt-sample/deployment.yaml"},
-		},
-		{
 			description: "multi manifest, mixed dir and file",
 			generateConfig: latest.Generate{
 				RawK8s: []string{"rawYaml-sample"},
-				Kustomize: &latest.Kustomize{
-					Paths: []string{"kustomize-sample"},
-				},
-				Kpt: []string{"kpt-sample"},
 			},
 			expected: []string{
-				"kpt-sample/Kptfile",
-				"kpt-sample/deployment.yaml",
 				"rawYaml-sample/pod.yaml",
 				"rawYaml-sample/pods2.yaml",
-				"kustomize-sample/kustomization.yaml",
-				"kustomize-sample/patch.yaml",
 			},
 		},
 	}
@@ -280,10 +168,6 @@ func TestManifestDeps(t *testing.T) {
 			tmpDir.Write("rawYaml-sample/pod.yaml", podYaml).
 				Write("rawYaml-sample/pods2.yaml", podsYaml).
 				Write("rawYaml-sample/irrelevant.txt", "").
-				Write("kustomize-sample/kustomization.yaml", kustomizeYaml).
-				Write("kustomize-sample/patch.yaml", patchYaml).
-				Write("kpt-sample/Kptfile", kptfileYaml).
-				Write("kpt-sample/deployment.yaml", kustomizeDeploymentYaml).
 				Touch("empty.ignored").
 				Chdir()
 			expectedPaths := []string{}
@@ -294,71 +178,6 @@ func TestManifestDeps(t *testing.T) {
 			actual, err := g.ManifestDeps()
 			t.CheckNoError(err)
 			t.CheckDeepEqual(expectedPaths, actual)
-		})
-	}
-}
-
-func TestBuildCommandArgs(t *testing.T) {
-	tests := []struct {
-		description   string
-		buildArgs     []string
-		kustomizePath string
-		expectedArgs  []string
-	}{
-		{
-			description:   "no BuildArgs, empty KustomizePaths ",
-			buildArgs:     []string{},
-			kustomizePath: "",
-			expectedArgs:  nil,
-		},
-		{
-			description:   "One BuildArg, empty KustomizePaths",
-			buildArgs:     []string{"--foo"},
-			kustomizePath: "",
-			expectedArgs:  []string{"--foo"},
-		},
-		{
-			description:   "no BuildArgs, non-empty KustomizePaths",
-			buildArgs:     []string{},
-			kustomizePath: "foo",
-			expectedArgs:  []string{"foo"},
-		},
-		{
-			description:   "One BuildArg, non-empty KustomizePaths",
-			buildArgs:     []string{"--foo"},
-			kustomizePath: "bar",
-			expectedArgs:  []string{"--foo", "bar"},
-		},
-		{
-			description:   "Multiple BuildArg, empty KustomizePaths",
-			buildArgs:     []string{"--foo", "--bar"},
-			kustomizePath: "",
-			expectedArgs:  []string{"--foo", "--bar"},
-		},
-		{
-			description:   "Multiple BuildArg with spaces, empty KustomizePaths",
-			buildArgs:     []string{"--foo bar", "--baz"},
-			kustomizePath: "",
-			expectedArgs:  []string{"--foo", "bar", "--baz"},
-		},
-		{
-			description:   "Multiple BuildArg with spaces, non-empty KustomizePaths",
-			buildArgs:     []string{"--foo bar", "--baz"},
-			kustomizePath: "barfoo",
-			expectedArgs:  []string{"--foo", "bar", "--baz", "barfoo"},
-		},
-		{
-			description:   "Multiple BuildArg no spaces, non-empty KustomizePaths",
-			buildArgs:     []string{"--foo", "bar", "--baz"},
-			kustomizePath: "barfoo",
-			expectedArgs:  []string{"--foo", "bar", "--baz", "barfoo"},
-		},
-	}
-
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			args := kustomizeBuildArgs(test.buildArgs, test.kustomizePath)
-			t.CheckDeepEqual(test.expectedArgs, args)
 		})
 	}
 }
